코드 정리&BMI
