--- users.txt
@@ -1,2 +1,0 @@
-user,password
-seoyun,tjdbs0191
